refactor: changed my mind on some of the refactopring for now
