+++ logwatcher/logwatcher.py
@@ -1,5 +1,5 @@
+#!/usr/local/bin/python2.7
 #!/bin/env python
-# !/usr/local/bin/python
 
 #  Copyright 2015 CityGrid Media, LLC
 #
@@ -23,6 +23,7 @@ import sys
 import atexit
 import ConfigParser
 import signal
+import argparse
 
 CODE_VERSION = "$Id: logwatcher.py 233274 2014-06-23 23:20:52Z heitritterw $"
 
@@ -103,7 +104,7 @@ class LogWatcher(object):
 
         if self.getPID() < 1:
             if self.daemonize:
-                self.procdaemonize()
+                procdaemonize()
 
         if self.lockPID() == 0:
             print "Pidfile found"
@@ -1092,36 +1093,156 @@ class LogWatcher(object):
             self.log_time += time.time() - self.log_time_start
 
 
-    def procdaemonize(self, stdin='/dev/null', stdout='/dev/null', stderr='/dev/null'):
-        # Do first fork.
-        try:
-            pid = os.fork()
-            if pid > 0:
-                sys.exit(0) # Exit first parent.
-        except OSError, e:
-            sys.stderr.write("fork #1 failed: (%d) %s\n" % (e.errno, e.strerror))
-            sys.exit(1)
+def procdaemonize(stdin='/dev/null', stdout='/dev/null', stderr='/dev/null'):
+    '''Daemonize this thing.'''
 
-        # Decouple from parent environment.
-        os.chdir("/")
-        os.umask(0)
-        os.setsid()
+    print "trying to daemonize"
+    # Do first fork.
+    try:
+        pid = os.fork()
+        print "fork 1"
+        if pid > 0:
+            sys.exit(0) # Exit first parent.
+    except OSError, e:
+        sys.stderr.write("fork #1 failed: (%d) %s\n" % (e.errno, e.strerror))
+        sys.exit(1)
 
-        # Do second fork.
-        try:
-            pid = os.fork()
-            if pid > 0:
-                sys.exit(0) # Exit second parent.
-        except OSError, e:
-            sys.stderr.write("fork #2 failed: (%d) %s\n" % (e.errno, e.strerror))
-            sys.exit(1)
+    # Decouple from parent environment.
+    os.chdir("/")
+    os.umask(0)
+    os.setsid()
 
-        # Now I am a daemon!
+    # Do second fork.
+    try:
+        pid = os.fork()
+        print "fork 2"
+        if pid > 0:
+            sys.exit(0) # Exit second parent.
+    except OSError, e:
+        sys.stderr.write("fork #2 failed: (%d) %s\n" % (e.errno, e.strerror))
+        sys.exit(1)
 
-        # Redirect standard file descriptors.
-        si = file(stdin, 'r')
-        so = file(stdout, 'a+')
-        se = file(stderr, 'a+', 0)
-        os.dup2(si.fileno(), sys.stdin.fileno())
-        os.dup2(so.fileno(), sys.stdout.fileno())
-        os.dup2(se.fileno(), sys.stderr.fileno())
+    # Now I am a daemon!
+    print "Now daemon"
+
+    # Redirect standard file descriptors.
+    si = file(stdin, 'r')
+    so = file(stdout, 'a+')
+    se = file(stderr, 'a+', 0)
+    os.dup2(si.fileno(), sys.stdin.fileno())
+    os.dup2(so.fileno(), sys.stdout.fileno())
+    os.dup2(se.fileno(), sys.stderr.fileno())
+    print "shouldn't see this"
+
+# FIXME: Why does this need a second arg?
+def handle_signal(signum, frame):
+    '''Exit on a given signal number.'''
+
+    print "\nLogWatcher killed"
+    sys.exit(signum)
+
+def parse_args():
+    '''Parse all the command line arguments.'''
+
+    help_desc = '''
+    Watch logs and send them to graphite.
+
+    >>> lw_daemon.py -D -V -c /app/logwatcher/conf/logwatcher.ini -b -g graphite
+    readlines()
+    '''
+
+    pap = argparse.ArgumentParser(formatter_class=argparse.RawTextHelpFormatter,
+                                  description=help_desc)
+
+    pap.add_argument('-b',
+                     '--beginning',
+                     action='store_true',
+                     help='Read the log from the beginning (useful with -q)')
+    pap.add_argument('-c',
+                     '--config',
+                     help='Use the given configuration file',
+                     default=None)
+    pap.add_argument('-d',
+                     '--daemonize',
+                     action='store_true',
+                     help="Run in the background.",
+                     default=0)
+    pap.add_argument('-D',
+                     '--debug',
+                     help="Don't send metrics, just print them.",
+                     default=0)
+    pap.add_argument('-G',
+                     '--use-graphite',
+                     dest='use_graphite',
+                     action='store_true',
+                     help='Use graphite, find server in /etc/graphite.conf.')
+    pap.add_argument('-g',
+                     '--graphite-server',
+                     help='Use graphite, with server GRAPHITE_SERVER.',
+                     default=None)
+    pap.add_argument('-i',
+                     '--distinguisher',
+                     help='Use the given string in the metric names',
+                     default=None)
+    pap.add_argument('-p',
+                     '--pidfile',
+                     help='Store the PID in the given file',
+                     default=None)
+    pap.add_argument('-R',
+                     '--prefix-root',
+                     dest='prefix_root',
+                     help='The prefix for all logwatcher metrics.',
+                     default='LW_')
+    pap.add_argument('-q',
+                     '--quit',
+                     action='store_true',
+                     help="Quit after sending metrics (useful with -D).")
+    pap.add_argument('-t',
+                     '--testconfig',
+                     action='store_true',
+                     help='Read overrides from the [test] section of the '
+                     'configuration file')
+    pap.add_argument('-V',
+                     '--verbose',
+                     action='store_true',
+                     help="Print gmetric commands as they're sent. Disables -D.")
+    pap.add_argument('-v',
+                     '--version',
+                     action='store_true',
+                     help="Print the version.")
+
+    return pap.parse_args()
+
+def main():
+    '''Do all the things.'''
+
+    # parse the args
+    args = parse_args()
+
+    # FIXME: convert to true/false or one option with levels.
+    # debug = 0
+    # for opt, arg in opts:
+    #     if opt in ("-D", "--debug"):
+    #         debug = 1
+    #     if opt in ("-V", "--verbose"):
+    #         debug = 2
+
+    if args.version:
+        print CODE_VERSION
+        sys.exit(0)
+
+    lw = LogWatcher(args.pidfile,
+                    args.daemonize,
+                    args.config,
+                    args.distinguisher,
+                    args.debug,
+                    args.quit,
+                    args.beginning,
+                    args.testconfig,
+                    args.graphite_server,
+                    args.use_graphite,
+                    args.prefix_root)
+
+if __name__ == "__main__":
+    signal.signal(signal.SIGINT, handle_signal)
+    main()
